Separated simulation package from simulation service
Before, the simulation package held the code for running a simulation as
well as the code needed for running the simulation service (docker). That is the
technical implementation of the simulation and the code to wire up the
simulation service with the message broker. With the new file structure
the latter does not belong into that package.
So I moved that code, the one to wire up the simulation to the message
broker, into its own service foler.

And with that I also started to introduce a new naming scheme: Packages
are named after objects, while services are named using personal nouns.