Server fixes
With this commit, the server works again.

In the `docker-compose.yml` I changed the command used to start the
server to reflect the new directory structure with its new import paths.

The `STATIC_DIR` variable is now correctly used in the servers code and
the static dir is mounted accordingly. The
`TEMPLATE_DIR` variable was removed from the values package. The
template dir is set relatively in the servers code as the templates live
in the servers service-folder.

I corrected an import that mistakenly referenced a package, while the
file to import was actually in the servers own service-folder.