Web-App: Separated code from config
In this commit I separated the web-apps code from its configuration.
This makes it easier to work with either of it as you have a better
overview. Also this structure is commonly used and one does not need to
adapt to it or learn anything new about it.

Changes were reflected in the `docker-compose.yml` as well as the
webpacks configuration. The bundler-service now outputs a bunlde into
the mounted static-volume which is used by the server-service as well.

The `static` folder was added to the `.gitignore` as it contains bundled
code which we do not want to commit.