[Julia] Adapted `Simulation` package structure
In this commit I adapted the `Simulation` package structure to match
Julias package structure. That is a `Manifest.toml` as well as a
`Project.toml` and all the code in a `src` folder with a file named
after the package inside of it. Just like I did with the `EventSystem`
in the previous commit.

I used `pkg.generate` to generate a new `Project.toml` and thus a new
uuid for this package to avoid conflicting uuids with other packages.
I also removed no longer needed dependencies.