[Julia] Adapted `Simulator` service structure
In this commit I adapted the `Simulator` - formerly `simulation` -
service structure to match Julias package structure. That is a
`Manifest.toml` as well as a `Project.toml` and all the code in a
`src` folder with a file named after the package inside of it. Just
like I did with the `EventSystem` and `Simulation` in the previous
commits.

The difference being here, that this is not a package, but a service.
Thus the personificated noun as name.

I did not find a way to wire up the needed local packages for this
service. This will be done in the next commits.