[Julia] Simulator service running again
With this commit I employed changes to make the julia simulator service
run again. It was broken because of the new service / package structure
I introduced in the last commits.

In the Simulator service I moved over code from the `worker.jl` file
into the `Simulator.jl` file as there is no need for this separation, it
just grew like this historically.
I added a `run` function to the service, so that it does not start
directly when executing `using Simulator`. This enables precompiling
the julia packages when building the docker image. (However there seems
to be a problem with a library called `Arpack` which makes the
precompilation fail anyway.)
Also the simulator now has the AMQPClient listed in its dependencies, as
it should have been right from the start. (This was added to the
`Project.toml` and also triggered the creation of the `Manifest.toml`)
I updated the services `Dockerfile` to work with the new structure. This
means copying all the code into it and then setting the packages in
development mode. This is a workaround for enabling the use of our
packages. Normally each package would need its own git repo. In the
future I hope to give each package its own repository.
I also adapted the `docker-compose.yml` to match the changes in the
`Dockerfile`. This is mainly the build context, so that the image can
copy in all the code when it gets build.

Besides that there are some small syntactical fixes.
Also I adapted the way we access environment variables, otherwise we end
up with an error when those are not provided - which would happen at
build time of the docker image.

Additional Notes:
The startup time of the Julia Simulator is horrendous! I hope to get
this fixed, when precompiling works again correctly and I have the
services and packages in their own repositories.