[Python] Fixed and renamed python services
In this commit I fixed and - since it was a good chance - also renamed the
python services. They are now named according to the unofficial naming
scheme only I have in my head so far. I'll put that down in our
guidelines when the restructuring is finished.

What did I do?
The `persistor` (formerly `dbWorker`), the `py_simulator` (formerly
`py_simulation`) and the server are all running again as they used to.
Changes include adapting the imports on the various services, not only
be cause of the renaming.
The `persistor` now is started with calling its `run` function, just
like the `py_simulator`.
I also needed to adapt the paths to the model JSON files and the
`static` dir as those changes with the structure. I adapted those paths
in some places already, but some I missed. Doing so I also improved the
error handling when the model JSON files get loaded.
I adapted the mounts and commands in the `docker-compose.yml` as well as
adapting the `Makefile`.