Makefile adaptations
with this commit I adapted the last bits and pieces in the `Makefile` for
the restructuring. This is to ensure the `Makefile` working properly
with the new structure and providing helpful information.

I made it so the `test` target does not continue running whenever
something fails. I think this makes it more obvious when a test fails,
as no other tests are run.

I changed the paths to be removed in the `clean` target, as those paths
changed with the restructuring.

I updated the `help` target, as I had failed to update it in the last
commits where I introduced changes to it.