Renamed environment variables to match service scheme
In this commit I renamed the last piece that was still using the old
naming scheme: The environment variables.
With this change, everything is now on the new naming scheme, which will
make it easier to distinguish between service and package.

I also adapted the `readme.md` accordingly.

Furthermore, I updated some now outdated info that changed with the
restructuring.