Python as the default simulator
With this commit I reimplemented the python simulator as the default
simulator service. This has no other reason than pleasing my ego. I
don't like to commit things that don't work, and currently the julia
simulation is not working and probably will require a lot of work to get
it working as we expect it to.