Extracted Simulation code
In this commit I removed all other code that does not belong to the
simulation. I wanted to have all previous commits, so that one can go
back (using git blame) if needed. But doing so all other code got into
this repo as well, which is why I removed it now. (Yeah, I know, it adds
some overhead when cloning, but I think this is bearable.)