Cleaned up tests after merge
With this commit I adapted the merged tests to the new structure of the
package. The tests now live in the `test` folder and can be executed via
the julia cli thanks to the `runtests.jl` file in the test folder.

I also moved the `simulationData.jl` into the src folder. It is not yet
used, but might serve a s a base for better code splitting which is a
thing that is missing in this repo.